// packages/schematic/nestjs/src/schematics/feature-microservice/schema.d.ts
export interface FeatureMicroserviceOptions {
  feature: string;
  dsn?: string;
  directory: string;
  imageRegistry?: string;
  imageName?: string;
  imageSuffix?: string;
  apiPrefix?: string;
  port?: number;
  randomPort?: boolean;
  apiConfigurationFile?: string;
  overwrite: boolean;
  openApi: boolean;
  jwt: boolean;
}